requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Limiter==3.3.0
python-socketio==5.8.0
python-engineio==4.7.1
SQLAlchemy==2.0.23
pytz==2023.3
pandas==2.1.3
openpyxl==3.1.2
XlsxWriter==3.1.9
qrcode==7.4.2
Werkzeug==2.3.7
schedule==1.2.0
email-validator==2.1.0


# Flask Framework


# Database
pymongo==4.6.0
dnspython==2.4.2

# Security


# Server
gunicorn==21.2.0
eventlet==0.33.3

# WebSocket
python-socketio==5.10.0

# Date/Time


# Utilities
python-dotenv==1.0.0

# Async support
gevent==23.9.1
gevent-websocket==0.10.1
